+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix accessibility barriers raised in the 2026-01 report:
   - 9.3.3.1 "Fehlererkennung/Allgemeine Fehlermeldung" (implement custom error messages for forms like "Please enter the publisher of this dataset.", rather than using generic ones like "Missing value").
 - Replace the generic `not_empty` validator function with a customizable one (`ckanext.berlin_dataset_schema.validation.not_empty_customizable`) to allow error messages that are specific to the metadata attribute that the function is used on.
+- Provide a more appropriate error message when no category is chosen in the UI.
 
 ## [0.3.13](https://github.com/berlinonline/ckanext-berlin_dataset_schema/releases/tag/0.3.13)